% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage-package.R
\docType{package}
\name{mypackage-package}
\alias{mypackage}
\alias{mypackage-package}
\title{mypackage: Halves-Ordered Variational Inference Tools}
\description{
Functions for simulation, batch fitting, and real-data analysis of
halves-ordered multivariate traits with fast variational inference.
}
\author{
\strong{Maintainer}: yytyyt11 Maintainer \email{819826397@qq.com}

}
\keyword{internal}
