% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialHVI-package.R
\docType{package}
\name{spatialHVI-package}
\alias{spatialHVI}
\alias{spatialHVI-package}
\title{spatialHVI: Spatial Heritability Variational Inference Tools}
\description{
Functions for simulation, batch fitting, and real-data analysis of
halves-ordered multivariate traits with fast variational inference.
Package maintenance contact: yytyyt11 Maintainer
\href{mailto:819826397@qq.com}{819826397@qq.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yytyyt11/mypackage}
  \item Report bugs at \url{https://github.com/yytyyt11/mypackage/issues}
}

}
\author{
\strong{Maintainer}: yytyyt11 Maintainer \email{819826397@qq.com}

}
\keyword{internal}
